Link source set classes to sourceSet.java classes
When updating the destinationDirectory of a JavaCompile task, the destination directory of the
SourceDirectorySet which is compiledBy it is not also updated. However, the classesDirectory of
the source directory set is updated to remain consistent with the compile task's output diretory.

Here, we update the SourceSet to reference the source directory set's classes directory so that
the source set outputs are kept consistent with the compile task's destination directory.

In order for this to work, we avoid registering source set outputs with build output cleanup registry
 since we already register the build directory as a whole