Prepare & cross the configuration time barrier in "vintage" builds
Given that the task execution time checks now depend on the
configuration time barrier, we need to manage its state not only in
configuration cache builds but in "vintage" ones as well.

Now that the "vintage" builds manage the barrier state, it is also
necessary to reset the barrier after the buildSrc tasks have executed,
so that the "main" build is configured in the barrier's configuration
time.